refactor: simplify `lib/record-scenario.js`
