+++ lib/record-scenario.js
@@ -1,7 +1,6 @@
 module.exports = recordScenario
 
 const nock = require('nock')
-const normalize = require('./normalize')
 const removeCredentials = require('./remove-credentials')
 
 async function recordScenario ({request, scenario}) {
@@ -33,27 +32,10 @@ async function recordScenario ({request, scenario}) {
     await scenario({request})
   }
 
-  const fixtures = {
-    raw: nock.recorder.play()
-  }
-
-  fixtures.raw.forEach(removeCredentials)
-  fixtures.normalized = fixtures.raw
-    .map(copy)
-    .filter(hasntIgnoreHeader)
-    .map(normalize)
+  const fixtures = nock.recorder.play()
 
   nock.recorder.clear()
   nock.restore()
 
-  return fixtures
-}
-
-function copy (object) {
-  return Object.assign({}, object)
-}
-
-function hasntIgnoreHeader (fixture) {
-  const hasIgnoreHeader = 'x-octokit-fixture-ignore' in fixture.reqheaders
-  return !hasIgnoreHeader
+  return fixtures.map(removeCredentials)
 }
